examples/research-assistant-demo/pyproject.toml
[project]
name = "research-assistant-demo"
version = "0.1.0"
description = "Research Assistant demo showcasing virtual tools for distributed microservices"
requires-python = ">=3.10"
dependencies = [
    # MCP services
    "mcp>=1.3.0",
    "pydantic>=2.0.0",
    "sentence-transformers>=2.2.0",
    "sqlite-vec>=0.1.1",
    "aiosqlite>=0.19.0",

    # HTTP client for external APIs
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",

    # Web UI
    "python-fasthtml>=0.4.0",
    "uvicorn>=0.30.0",

    # Agent framework
    "google-adk>=0.5.0",

    # LLM providers (configurable)
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",

    # Utilities
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
